Simplify code per code reviews
